+++ packages/flagship/android/app/src/main/java/com/brandingbrand/reactnative/and/flagship/MainApplication.java
@@ -4,8 +4,8 @@ import android.app.Application;
 import android.content.Context;
 import com.facebook.react.PackageList;
 import com.facebook.react.ReactApplication;
+import com.facebook.react.ReactInstanceManager;
 import com.facebook.react.ReactNativeHost;
-import com.avishayil.rnrestart.ReactNativeRestartPackage;
 import com.facebook.react.ReactPackage;
 import com.facebook.soloader.SoLoader;
 import java.lang.reflect.InvocationTargetException;
@@ -15,39 +15,38 @@ import com.reactnativenavigation.react.NavigationReactNativeHost;
 import com.reactnativenavigation.react.ReactGateway;
 
 import java.util.Arrays;
+import java.util.ArrayList;
 import java.util.List;
 
 public class MainApplication extends NavigationApplication {
-    protected List<ReactPackage> getPackages(Application application, boolean debug) {
-      @SuppressWarnings("UnnecessaryLocalVariable")
-      List<ReactPackage> packages = new PackageList(this).getPackages();
-      // Packages that cannot be autolinked yet can be added manually here, for example:
-      packages.add(new NativeConstantsPackage());
-      packages.add(new EnvSwitcherPackage());
-      packages.add(new ReactNativeRestartPackage());
-      return packages;
-    }
-
     @Override
     protected ReactGateway createReactGateway() {
-        ReactNativeHost host = new NavigationReactNativeHost(this, isDebug(), createAdditionalReactPackages()) {
+        ReactNativeHost host = new NavigationReactNativeHost(this) {
+            @Override
             protected String getJSMainModuleName() {
                 return "index";
             }
+            @Override
+            public boolean getUseDeveloperSupport() {
+                return BuildConfig.DEBUG;
+            }
+            @Override
+            public List<ReactPackage> getPackages() {
+                List<ReactPackage> packages = new PackageList(this).getPackages();
+                // Packages that cannot be autolinked yet can be added manually here, for example:
+                packages.add(new NativeConstantsPackage());
+                packages.add(new EnvSwitcherPackage());
+                return packages;
+            }
+
             // [CODEPUSH FUNCTIONS INJECT]
         };
-        return new ReactGateway(this, isDebug(), host);
+        return new ReactGateway(host);
     }
 
-    @Override
     public boolean isDebug() {
         // Make sure you are using BuildConfig from your own application
         return BuildConfig.DEBUG;
-    }
-
-    @Override
-    public List<ReactPackage> createAdditionalReactPackages() {
-        return getPackages(this, BuildConfig.DEBUG);
     }
 
     public String getJSMainModuleName() {
@@ -85,23 +84,28 @@ public class MainApplication extends NavigationApplication {
     public void onCreate() {
         super.onCreate();
         SoLoader.init(this, /* native exopackage */ false);
-        initializeFlipper(this); // Remove this line if you don't want Flipper enabled
+        initializeFlipper(this, getReactNativeHost().getReactInstanceManager());
     }
 
     /**
-     * Loads Flipper in React Native templates.
-     *
-     * @param context
-     */
-    private static void initializeFlipper(Context context) {
-        if (BuildConfig.DEBUG) {
-            try {
-                /*
-                We use reflection here to pick up the class that initializes Flipper,
-                since Flipper library is not available in release mode
-                */
-                Class<?> aClass = Class.forName("com.facebook.flipper.ReactNativeFlipper");
-                aClass.getMethod("initializeFlipper", Context.class).invoke(null, context);
+   * Loads Flipper in React Native templates. Call this in the onCreate method with something like
+   * initializeFlipper(this, getReactNativeHost().getReactInstanceManager());
+   *
+   * @param context
+   * @param reactInstanceManager
+   */
+  private static void initializeFlipper(
+      Context context, ReactInstanceManager reactInstanceManager) {
+    if (BuildConfig.DEBUG) {
+      try {
+        /*
+         We use reflection here to pick up the class that initializes Flipper,
+        since Flipper library is not available in release mode
+        */
+        Class<?> aClass = Class.forName("CONFIG_BUNDLE_ID.ReactNativeFlipper");
+        aClass
+            .getMethod("initializeFlipper", Context.class, ReactInstanceManager.class)
+            .invoke(null, context, reactInstanceManager);
             } catch (ClassNotFoundException e) {
                 e.printStackTrace();
             } catch (NoSuchMethodException e) {