+++ packages/fsi18n/src/translations/en.ts
@@ -335,6 +335,9 @@ export const keys: FSTranslationKeys = {
     emailForm: {
       placeholder: 'Email',
       error: 'Required Field'
+    },
+    button: {
+      apply: 'Apply'
     }
   }
 };