+++ packages/fsproductindex/src/components/ProductIndexGrid.tsx
@@ -21,19 +21,36 @@ import {
 
 import { style as S } from '../styles/ProductIndex';
 import {
+  Image,
   ListRenderItemInfo,
+  SafeAreaView,
   StyleProp,
+  StyleSheet,
   Text,
   TouchableOpacity,
   View,
   ViewStyle
 } from 'react-native';
 import FSI18n, { translationKeys } from '@brandingbrand/fsi18n';
+
+const images = {
+  close: require('../../assets/images/iconClose.png')
+};
+
+const styles = StyleSheet.create({
+  cancelButton: {
+    position: 'absolute',
+    top: 16,
+    right: 21.5
+  }
+});
+
 const componentTranslationKeys = translationKeys.flagship.productIndex;
 
 export interface PropTyps extends ProductIndexPropType {
   onPress: (data: CommerceTypes.Product) => () => void;
   containerStyle?: StyleProp<ViewStyle>;
+  headerWithCancelButton?: boolean;
 }
 
 const defaultErrorMessage =
@@ -41,11 +58,10 @@ const defaultErrorMessage =
 const SORT_ITEM_KEY = '__pirate_sort';
 
 export interface StateType {
-  sortModalVisble: boolean;
-  filterModalVisble: boolean;
+  sortModalVisible: boolean;
+  filterModalVisible: boolean;
   isLoading: boolean;
   isMoreLoading: boolean;
-  hasAnotherPage: boolean;
   hasFetchError: boolean;
 }
 
@@ -53,38 +69,54 @@ export default class ProductIndexGrid extends Component<
   PropTyps & WithProductIndexProps,
   StateType
 > {
-  page: number = 1;
-
   constructor(props: PropTyps & WithProductIndexProps) {
     super(props);
 
     const { commerceData, onLoadComplete } = props;
-
-    const hasAnotherPage = this.hasAnotherPage(commerceData);
+    let maxPageLoaded = 1;
+    let maxCount = 1;
+    if (commerceData) {
+      maxCount = this.maxCount(commerceData);
+      if (commerceData.page) {
+        maxPageLoaded = commerceData.page;
+        if (commerceData.limit && commerceData.products) {
+          maxCount = (commerceData.limit * (commerceData.page - 1)) + commerceData.products.length;
+        }
+      }
+      if (onLoadComplete) {
+        onLoadComplete(this.loadMore, maxPageLoaded < this.maxPage(commerceData),
+          maxCount, maxCount);
+      }
+    }
 
     this.state = {
-      sortModalVisble: false,
-      filterModalVisble: false,
+      sortModalVisible: false,
+      filterModalVisible: false,
       isLoading: false,
       isMoreLoading: false,
-      hasFetchError: false,
-      hasAnotherPage
+      hasFetchError: false
     };
 
-    if (commerceData && onLoadComplete) {
-      const count = commerceData.products && commerceData.products.length;
+  }
 
-      onLoadComplete(this.loadMore, hasAnotherPage, count, count);
+  maxCount = (commerceData?: CommerceTypes.ProductIndex): number => {
+    if (commerceData && commerceData.limit && commerceData.total && commerceData.page) {
+      const maxPage = this.maxPage(commerceData);
+      if (commerceData.page < maxPage) {
+        return commerceData.page * commerceData.limit;
+      } else {
+        return commerceData.total;
+      }
+    } else {
+      return 0;
     }
   }
 
-  hasAnotherPage = (data?: CommerceTypes.ProductIndex) => {
-    if (!data || !data.page || !data.total) {
-      return false;
+  maxPage = (commerceData?: CommerceTypes.ProductIndex) => {
+    if (commerceData && commerceData.total && commerceData.limit) {
+      return Math.ceil(commerceData.total / commerceData.limit);
     }
-
-    // fall back to count instead of limit in instances where limit wasn't specified in query
-    return data.page * (data.limit || data.products.length) < data.total;
+    return 1;
   }
 
   renderItem = ({ item }: ListRenderItemInfo<CommerceTypes.Product>): JSX.Element => {
@@ -96,6 +128,7 @@ export default class ProductIndexGrid extends Component<
     return (
       <ProductItem
         style={S.productItem}
+        id={item.id}
         title={item.title}
         brand={item.brand}
         image={item.images && item.images.find(img => !!img.uri)}
@@ -112,7 +145,7 @@ export default class ProductIndexGrid extends Component<
     );
   }
 
-  renderHeader = () => {
+  renderActionBar = () => {
     const {
       commerceData,
       hideActionBar,
@@ -129,7 +162,9 @@ export default class ProductIndexGrid extends Component<
       return renderRefineActionBar(
         this.showFilterModal,
         this.showSortModal,
-        commerceData
+        commerceData,
+        this.handleFilterApply,
+        this.handleFilterReset
       );
     }
 
@@ -144,23 +179,60 @@ export default class ProductIndexGrid extends Component<
     );
   }
 
+  renderHeader = () => {
+    const { commerceData } = this.props;
+
+    if (!commerceData) {
+      return null;
+    }
+
+    let loadPrev: JSX.Element | null = null;
+    if (this.props.renderLoadPrev) {
+      if (this.state.isMoreLoading) {
+        loadPrev = this.props.renderLoading ? (
+          this.props.renderLoading()
+        ) : (
+          <Loading
+            style={[
+              S.loading,
+              S.loadingLoadMore,
+              this.props.loadMoreLoadingStyle
+            ]}
+          />
+        );
+      } else {
+        loadPrev = this.props.renderLoadPrev(
+          this.loadPrev,
+          (commerceData.minPage || commerceData.page || 1) > 1
+        );
+      }
+    }
+
+    return (
+      <View>
+        {loadPrev}
+        {this.renderActionBar()}
+      </View>
+    );
+  }
+
   showFilterModal = () => {
-    this.setState({ filterModalVisble: true });
+    this.setState({ filterModalVisible: true });
   }
 
   closeFilterModal = () => {
-    this.setState({ filterModalVisble: false });
+    this.setState({ filterModalVisible: false });
   }
 
   showSortModal = () => {
-    this.setState({ sortModalVisble: true });
+    this.setState({ sortModalVisible: true });
   }
 
   closeSortModal = () => {
-    this.setState({ sortModalVisble: false });
+    this.setState({ sortModalVisible: false });
   }
 
-  handleFilterApply = (selectedItems: any, info: any) => {
+  handleFilterApply = (selectedItems: any, info?: { isButtonPress: boolean }) => {
     if (!this.props.filterInBackground) {
       this.closeFilterModal();
     } else {
@@ -182,7 +254,7 @@ export default class ProductIndexGrid extends Component<
     }
 
     if (this.props.handleFilterApply) {
-      this.props.handleFilterApply(selectedItems);
+      this.props.handleFilterApply(selectedItems, info);
     } else {
       this.reloadByQuery({
         ...refinementsQuery,
@@ -202,13 +274,25 @@ export default class ProductIndexGrid extends Component<
     }
   }
 
-  handleSortChange = (sortItem: CommerceTypes.SortingOption) => {
+  handleSortChange = (selectedItems: any) => (sortItem: CommerceTypes.SortingOption) => {
+    let refinementsQuery = {};
+
+    if (Object.keys(selectedItems).length > 0) {
+      refinementsQuery = {refinements: selectedItems};
+    }
+
     this.closeSortModal();
     if (this.props.handleSortChange) {
       this.props.handleSortChange(sortItem.id);
+    } else if (sortItem.id === 'default') {
+      this.reloadByQuery({
+        sortBy: undefined,
+        ...refinementsQuery
+      });
     } else {
       this.reloadByQuery({
-        sortBy: sortItem.id
+        sortBy: sortItem.id,
+        ...refinementsQuery
       });
     }
   }
@@ -216,7 +300,12 @@ export default class ProductIndexGrid extends Component<
   reloadByQuery = (query: CommerceTypes.ProductQuery) => {
     this.setState({ isLoading: true, hasFetchError: false });
     this.fetchByExtraQuery(query)
-      .then(this.handleNewData)
+      .then((data: CommerceTypes.ProductIndex) => {
+        this.handleNewData(data);
+        this.setState({
+          isLoading: false
+        });
+      })
       .catch(() => {
         this.setState({
           isLoading: false,
@@ -248,6 +337,9 @@ export default class ProductIndexGrid extends Component<
           this.closeFilterModal();
         }
         this.handleNewData(data);
+        this.setState({
+          isLoading: false
+        });
       })
       .catch(() => {
         this.setState({
@@ -257,17 +349,16 @@ export default class ProductIndexGrid extends Component<
       });
   }
 
-  handleNewData = (data: any) => {
-    this.page = 1;
-    const hasAnotherPage = this.hasAnotherPage(data);
+  handleNewData = (data: CommerceTypes.ProductIndex) => {
+    const newState: any = {};
+    const maxPageLoaded = data.page || 1;
+    const maxCount = this.maxCount(data);
+
     if (this.props.onLoadComplete) {
-      const count = data.products && data.products.length;
-      this.props.onLoadComplete(this.loadMore, hasAnotherPage, count, count);
+      this.props.onLoadComplete(this.loadMore, maxPageLoaded < this.maxPage(data),
+        maxCount, maxCount);
     }
-    this.setState({
-      isLoading: false,
-      hasAnotherPage
-    });
+    this.setState(newState);
 
     if (this.props.commerceLoadData) {
       this.props.commerceLoadData(data);
@@ -326,6 +417,31 @@ export default class ProductIndexGrid extends Component<
     return newQuery;
   }
 
+  renderCancelButton = (onPress: () => void) => {
+    return (
+      <TouchableOpacity
+        style={styles.cancelButton}
+        onPress={onPress}
+        activeOpacity={0.7}
+      >
+        <Image source={images.close} />
+      </TouchableOpacity>
+    );
+  }
+
+  renderCancelText = (onPress: () => void) => {
+    return (
+      <TouchableOpacity
+        style={[S.modalHeaderClose, this.props.modalCancelStyle]}
+        onPress={onPress}
+      >
+        <Text style={S.modalHeaderCloseText}>
+          {FSI18n.string(componentTranslationKeys.cancel)}
+        </Text>
+      </TouchableOpacity>
+    );
+  }
+
   renderModalHeader = ({ title, onPress }: any) => {
     const drilldownStyle =
       this.props.filterType === 'drilldown'
@@ -336,14 +452,11 @@ export default class ProductIndexGrid extends Component<
       <View
         style={[S.modalHeader, drilldownStyle, this.props.modalHeaderStyle]}
       >
-        <TouchableOpacity
-          style={[S.modalHeaderClose, this.props.modalCancelStyle]}
-          onPress={onPress}
-        >
-          <Text style={S.modalHeaderCloseText}>
-            {FSI18n.string(componentTranslationKeys.cancel)}
-          </Text>
-        </TouchableOpacity>
+        {
+          this.props.headerWithCancelButton
+            ? this.renderCancelButton(onPress)
+            : this.renderCancelText(onPress)
+        }
         <Text style={[S.modalHeaderText, this.props.modalHeaderTextStyle]}>
           {title}
         </Text>
@@ -351,38 +464,58 @@ export default class ProductIndexGrid extends Component<
     );
   }
 
+  // tslint:disable-next-line:cyclomatic-complexity
   renderSortModal = () => {
     if (this.props.hideActionBar) {
       return null;
     }
 
-    const { commerceData } = this.props;
+    const { commerceData, defaultSortOption } = this.props;
     let content = null;
+
+    const sortOptions = commerceData?.sortingOptions ? [...commerceData.sortingOptions] : [];
+
+    if (defaultSortOption) {
+      sortOptions.unshift({
+        id: 'default',
+        title: defaultSortOption
+      });
+    }
+
+    const selectedOption = commerceData?.selectedSortingOption ?
+      commerceData?.selectedSortingOption : 'default';
+
+    const selectedItems = this.props.mergeSortToFilter && commerceData?.selectedSortingOption
+      ? this.mergeSelectedRefinementsAndSort(
+        commerceData.selectedRefinements,
+        commerceData.selectedSortingOption
+      )
+      : commerceData?.selectedRefinements;
 
     if (commerceData) {
       if (this.props.renderSort) {
-        content = this.props.renderSort(this.handleSortChange, commerceData);
-      } else {
+        content = this.props.renderSort(this.handleSortChange(selectedItems), commerceData);
+      } else if (sortOptions) {
         content = (
-          <View style={S.modalContainer}>
+          <SafeAreaView style={S.modalContainer}>
             {this.renderModalHeader({
               title: this.props.sortHeaderStyle || 'Sort By',
               onPress: this.closeSortModal
             })}
             <SelectableList
-              items={commerceData.sortingOptions}
-              onChange={this.handleSortChange}
-              selectedId={commerceData.selectedSortingOption}
+              items={sortOptions}
+              onChange={this.handleSortChange(selectedItems)}
+              selectedId={selectedOption}
               {...this.props.sortListProps}
             />
-          </View>
+          </SafeAreaView>
         );
       }
     }
 
     return this.renderModal({
       content,
-      visible: this.state.sortModalVisble,
+      visible: this.state.sortModalVisible,
       closeModal: this.closeSortModal
     });
   }
@@ -471,21 +604,22 @@ export default class ProductIndexGrid extends Component<
         : commerceData.selectedRefinements;
 
       content = (
-        <View style={S.modalContainer}>
+        <SafeAreaView style={S.modalContainer}>
           <FilterListDrilldown
             items={items}
             onApply={this.handleFilterApply}
             onReset={this.handleFilterReset}
+            onClose={this.props.showDrilldownClose ? this.closeFilterModal : undefined}
             selectedItems={selectedItems}
             renderFilterItem={this.renderItemForCombinedFilterAndSort}
             renderFilterItemValue={this.renderItemValueForCombinedFilterAndSort}
             applyOnSelect={this.props.filterInBackground}
             singleFilterIds={
-              this.props.mergeSortToFilter ? [SORT_ITEM_KEY] : null
+              this.props.mergeSortToFilter ? [SORT_ITEM_KEY] : undefined
             }
             {...this.props.FilterListDrilldownProps}
           />
-        </View>
+        </SafeAreaView>
       );
     } else {
       content = (
@@ -496,7 +630,7 @@ export default class ProductIndexGrid extends Component<
           })}
 
           <FilterList
-            items={commerceData.refinements}
+            items={commerceData.refinements || []}
             onApply={this.handleFilterApply}
             onReset={this.handleFilterReset}
             selectedItems={commerceData.selectedRefinements}
@@ -508,7 +642,7 @@ export default class ProductIndexGrid extends Component<
 
     return this.renderModal({
       content,
-      visible: this.state.filterModalVisble,
+      visible: this.state.filterModalVisible,
       closeModal: this.closeFilterModal
     });
   }
@@ -553,7 +687,7 @@ export default class ProductIndexGrid extends Component<
           key={index}
           title={value.title}
           selected={selected}
-          onPress={this.handleSortSelectedInRefine(value)}
+          onPress={this.handleSortSelectedInRefine(value, selected)}
           {...selectableRowProps}
         />
       );
@@ -562,9 +696,10 @@ export default class ProductIndexGrid extends Component<
     }
   }
 
-  handleSortSelectedInRefine = (value: any) => () => {
+  handleSortSelectedInRefine = (value: any, selected: any) => () => {
     this.closeFilterModal();
-    this.handleSortChange(value);
+    // TODO: Test is needed.
+    this.handleSortChange(selected)(value);
   }
 
   renderNoResult = () => {
@@ -605,13 +740,13 @@ export default class ProductIndexGrid extends Component<
     );
   }
 
-  loadMore = () => {
+  loadPage = (page: number) => {
     const {
       commerceData,
       commerceProviderLoadMore
     } = this.props;
 
-    if (!commerceData || !commerceData.page) {
+    if (!commerceData) {
       // Cannot load more
       return;
     }
@@ -620,29 +755,13 @@ export default class ProductIndexGrid extends Component<
       isMoreLoading: true
     });
 
-    const newQuery = this.newProductQuery({ page: commerceData.page + 1 });
+    const newQuery = this.newProductQuery({ page });
     if (commerceProviderLoadMore) {
       commerceProviderLoadMore(newQuery)
-        .then(data => {
-          const hasAnotherPage = this.hasAnotherPage(data);
-          let totalCount: number = 0;
-
-          // TODO: Pageable properties should not be optional on Product Index type
-          if (data.limit && data.page) {
-            totalCount = (data.limit * (data.page - 1)) + data.products.length;
-          }
-
-          if (this.props.onLoadComplete) {
-            this.props.onLoadComplete(
-              this.loadMore,
-              hasAnotherPage,
-              totalCount,
-              data.products.length
-            );
-          }
+        .then((data: CommerceTypes.ProductIndex) => {
+          this.handleNewData(data);
           this.setState({
-            isMoreLoading: false,
-            hasAnotherPage
+            isMoreLoading: false
           });
         })
         .catch(() => {
@@ -653,7 +772,24 @@ export default class ProductIndexGrid extends Component<
     }
   }
 
+  loadPrev = () => {
+    const { commerceData } = this.props;
+    if (commerceData) {
+      this.loadPage((commerceData.minPage || commerceData.page || 1) - 1);
+    }
+  }
+
+  loadMore = () => {
+    const { commerceData } = this.props;
+    if (commerceData) {
+      this.loadPage((commerceData.page || 1) + 1);
+    }
+  }
+
   renderFooter = () => {
+    const { commerceData } = this.props;
+    const hasAnotherPage: boolean = commerceData !== undefined && commerceData.page !== undefined &&
+      commerceData.page < this.maxPage(commerceData);
     if (this.state.isMoreLoading) {
       return this.props.renderLoading ? (
         this.props.renderLoading()
@@ -671,11 +807,11 @@ export default class ProductIndexGrid extends Component<
     if (this.props.renderLoadMore) {
       return this.props.renderLoadMore(
         this.loadMore,
-        this.state.hasAnotherPage
+        hasAnotherPage
       );
     }
 
-    if (!this.state.hasAnotherPage) {
+    if (!hasAnotherPage) {
       return null;
     }
 
@@ -693,6 +829,7 @@ export default class ProductIndexGrid extends Component<
     );
   }
 
+  // tslint:disable-next-line: cyclomatic-complexity
   render(): React.ReactNode {
     const {
       commerceData,
@@ -706,6 +843,9 @@ export default class ProductIndexGrid extends Component<
     } = this.props;
 
     if (this.state.isLoading && !this.props.filterInBackground) {
+      if (this.props.renderGhost) {
+        return this.props.renderGhost();
+      }
       return <Loading style={[S.loading, loadingStyle]} />;
     }
 
@@ -732,8 +872,8 @@ export default class ProductIndexGrid extends Component<
           renderFooter={this.renderFooter}
           gridProps={gridProps}
         />
-        {this.state.sortModalVisble && this.renderSortModal()}
-        {this.state.filterModalVisble && this.renderFilterModal()}
+        {this.state.sortModalVisible && this.renderSortModal()}
+        {this.state.filterModalVisible && this.renderFilterModal()}
       </View>
     );
   }