+++ packages/pirateship/src/lib/datasource.ts
@@ -11,7 +11,9 @@ let reviewDataSource: ReviewDataSource;
 export interface DataSourceConfig {
   type: 'bbplatform' | 'commercecloud' | 'shopify' | 'mock';
   categoryFormat: 'grid' | 'list';
-  apiConfig: any;
+  apiConfig: {
+    apiHost: string;
+  };
   enableProxy?: boolean;
 }